requirements.txt
streamlit
geopandas
matplotlib
numpy
folium
pyproj
Pillow
shapely
streamlit-folium
rasterio